Adapted first block to second round: 4 sessions structure
