+++ intermediate/exercises/mod_01/tests.py
@@ -1,10 +1,10 @@
 #-*- coding: utf-8 -*-
 u'''
-Test MOD 15: mutable and immutable types common errors
+Test MOD 01: mutable and immutable types common errors
 '''
 import unittest
-import exercise_0 as source
-# import solution_0 as source
+import exercise as source
+# import solution as source
 
 
 class VerboseTestCase(unittest.TestCase):
